Command manager abstraction
Command manager is a bad name, but I don't have a better one currently. This is basically "the other half" of Window which is responsible for actual commands we execute on the GPU.
When this is used in conjunction with the Window class, we get a simple draw loop: get a command buffer, acquire swapchain image, submit the command buffer, present the swapchain image.
All GPU-CPU and GPU-GPU frame sync logic is now handled by these two classes for simple applications.